requirements.txt
streamlit
boto3
faiss-cpu
numpy
PyPDF2
requests
sentencepiece
cohere
mistralai